restructuring - see description
The aim for this branch is to allow more customization of the paths.
Prior to this, the path in the actual file system would have to be
“assets/js/foo.js” and the obfuscated url was always something like
assets/?asset=foo.js

This branch also removes the query string and gives customization of
the default asset root folder of “assets” to something else.

I’m also lowering the version because this was always in dev and
probably has a long way to go.